Create 1695_max_erasure_value.cpp
We use the prefix sum vector to calculate values in O(1). 
We initialise our start index and see if there's a duplicate. If a duplicate is not there then simply add the value to the current sum.
If a duplicate found, compare the answer with the current sum and use prefix sum vector to calculate current sum from the next index of the duplicated element.